feat: :construction: WIP - Size and export fixes
- Fix export error for projection tiffstacks
- Use float32 for projections and reconstructions
- Use courser estimations for file sizes (they're guesses anyways)

+++ webct/components/sim/clients/SimClient.py
@@ -401,7 +401,7 @@ class SimClient(Process):
 
 			# return result
 			print("Simulation completed.")
-			return result
+			return result.astype(np.float32)
 		else:
 			raise SimThreadError(
 				f"Unexpected response: {response}, wanted SimResponse.DONE"
@@ -460,7 +460,7 @@ class SimClient(Process):
 
 			# return result
 			print("Simulation completed.")
-			return result
+			return result.astype(np.float32)
 		else:
 			raise SimThreadError(
 				f"Unexpected response: {response}, wanted SimResponse.DONE"
